+++ NEWS.md
@@ -1,3 +1,4 @@
+* Single-item list of data frames can be exported fix #385
 * Move `stringi` to Suggests to reduce compilation time. Add an attribution to the internal data to list out all required packages #378
 * Move `readr` to Imports for `fwf`. `readr` is a dependency of `haven` so it does not increase the number of dependencies. Remove the original `read.fwf2` which doesn't guess `widths`. Keep the `widths` and `col.names` to maintain compatibility. #381